+++ packages/plugin-service-base/CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Changed
+
+- Added package metadata and limit files included in packages. Ensure `*.tsbuildinfo` files are absent to greatly shink package size. [[#177](https://github.com/Shopify/sewing-kit-next/pull/177)]
+- Remove unused dependencies on `@sewing-kit/hooks`, `@sewing-kit/plugin-javascript` and `koa-webpack` [[#177](https://github.com/Shopify/sewing-kit-next/pull/177)]
 
 ## 0.2.2 - 2021-06-18